refactor: rename core utilities module
